# server/api-documentation.yaml
openapi: 3.0.0
info:
  title: Super App API
  description: API documentation for the Superapp project.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /products:
    post:
      tags:
        - Products
      summary: Create a product for a business
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                short_description:
                  type: string
                description:
                  type: string
                picture:
                  type: string
                price_before:
                  type: number
                price_after:
                  type: number
                businessId:
                  type: string
      responses:
        "200":
          description: Product created successfully
  /products/{businessId}:
    get:
      tags:
        - Products
      summary: Get all products for a business
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of products
  /products/{id}:
    put:
      tags:
        - Products
      summary: Update a product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                short_description:
                  type: string
                description:
                  type: string
                picture:
                  type: string
                price_before:
                  type: number
                price_after:
                  type: number
      responses:
        "200":
          description: Product updated successfully
    delete:
      tags:
        - Products
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product removed successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT